requirements.txt
pyside6
sounddevice
vosk
keyboard
deep-translator
SpeechRecognition
openai-whisper
numpy
torch
webrtcvad
pydantic>=2.0.0
pytest>=7.0.0
pytest-qt>=4.0.0
